Moved classes to appropriate namespace
