+++ Api/Infrastructure/SupplierConnectors/ConnectorClient.cs
@@ -15,7 +15,7 @@ using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using Polly;
 
-namespace HappyTravel.Edo.Api.Infrastructure.DataProviders
+namespace HappyTravel.Edo.Api.Infrastructure.SupplierConnectors
 {
     public class ConnectorClient : IConnectorClient
     {
