+++ Api/Infrastructure/SupplierConnectors/IConnectorSecurityTokenManager.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace HappyTravel.Edo.Api.Infrastructure.DataProviders
+namespace HappyTravel.Edo.Api.Infrastructure.SupplierConnectors
 {
     public interface IConnectorSecurityTokenManager
     {
