+++ Api/Infrastructure/VoidObject.cs
@@ -1,4 +1,4 @@
-namespace HappyTravel.Edo.Api.Infrastructure.DataProviders
+namespace HappyTravel.Edo.Api.Infrastructure
 {
     /// <summary>
     ///     Empty object, used for Result T,E  to provide typed error with no result.
