+++ Api/Services/Accommodations/Bookings/IBookingManagementService.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
-using HappyTravel.Edo.Api.Infrastructure.DataProviders;
+using HappyTravel.Edo.Api.Infrastructure;
 using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Data.Management;
 using Microsoft.AspNetCore.Mvc;
